+++ azure-quantum/tests/unit/aio/test_aio_job.py
@@ -91,7 +91,6 @@ class TestJob(QuantumTestBase):
     Tests the azure.quantum.job module.
     """
 
-
     mock_create_job_id_name = "create_job_id"
     create_job_id = Job.create_job_id
