+++ azure-quantum/tests/unit/recordings/test_job_submit_quantinuum.yaml
@@ -61,7 +61,7 @@ interactions:
         "access_token": "fake_token"}'
     headers:
       content-length:
-      - '227'
+      - '223'
       content-type:
       - application/json; charset=utf-8
       transfer-encoding:
@@ -81,7 +81,7 @@ interactions:
       User-Agent:
       - azsdk-python-storage-blob/12.9.0 Python/3.7.10 (Darwin-21.3.0-x86_64-i386-64bit)
       x-ms-date:
-      - Fri, 18 Feb 2022 22:54:41 GMT
+      - Sat, 19 Feb 2022 01:14:54 GMT
       x-ms-version:
       - '2020-10-02'
     method: GET
@@ -89,7 +89,7 @@ interactions:
   response:
     body:
       string: "\uFEFF<?xml version=\"1.0\" encoding=\"utf-8\"?><Error><Code>ContainerNotFound</Code><Message>The
-        specified container does not exist.\nRequestId:4cba80da-901e-0041-421a-257d4d000000\nTime:2022-02-18T22:54:41.9753181Z</Message></Error>"
+        specified container does not exist.\nRequestId:fa23b072-d01e-006f-382e-252f5a000000\nTime:2022-02-19T01:14:55.3139367Z</Message></Error>"
     headers:
       content-length:
       - '225'
@@ -114,7 +114,7 @@ interactions:
       User-Agent:
       - azsdk-python-storage-blob/12.9.0 Python/3.7.10 (Darwin-21.3.0-x86_64-i386-64bit)
       x-ms-date:
-      - Fri, 18 Feb 2022 22:54:41 GMT
+      - Sat, 19 Feb 2022 01:14:55 GMT
       x-ms-version:
       - '2020-10-02'
     method: PUT
@@ -142,7 +142,7 @@ interactions:
       User-Agent:
       - azsdk-python-storage-blob/12.9.0 Python/3.7.10 (Darwin-21.3.0-x86_64-i386-64bit)
       x-ms-date:
-      - Fri, 18 Feb 2022 22:54:42 GMT
+      - Sat, 19 Feb 2022 01:14:55 GMT
       x-ms-version:
       - '2020-10-02'
     method: GET
@@ -185,7 +185,7 @@ interactions:
       x-ms-blob-type:
       - BlockBlob
       x-ms-date:
-      - Fri, 18 Feb 2022 22:54:42 GMT
+      - Sat, 19 Feb 2022 01:14:55 GMT
       x-ms-version:
       - '2020-10-02'
     method: PUT
@@ -215,7 +215,7 @@ interactions:
       Connection:
       - keep-alive
       Content-Length:
-      - '615'
+      - '611'
       Content-Type:
       - application/json
       User-Agent:
@@ -231,13 +231,13 @@ interactions:
         "quantinuum-job", "id": "00000000-0000-0000-0000-000000000000", "status":
         "Waiting", "outputDataFormat": "honeywell.quantum-results.v1", "outputDataUri":
         "https://st8bd7d3874f6b42beb1d90a.blob.core.windows.net:443/job-00000000-0000-0000-0000-000000000000/outputData?sv=PLACEHOLDER&sig=PLACEHOLDER&se=PLACEHOLDER&srt=co&ss=b&sp=racw",
-        "creationTime": "2022-02-18T22:54:42.9672408+00:00", "beginExecutionTime":
+        "creationTime": "2022-02-19T01:14:56.3180959+00:00", "beginExecutionTime":
         null, "endExecutionTime": null, "cancellationTime": null, "costEstimate":
         null, "errorData": null, "isCancelling": false, "tags": [], "access_token":
         "fake_token"}'
     headers:
       content-length:
-      - '1181'
+      - '1175'
       content-type:
       - application/json; charset=utf-8
       transfer-encoding:
@@ -267,13 +267,13 @@ interactions:
         "quantinuum-job", "id": "00000000-0000-0000-0000-000000000000", "status":
         "Succeeded", "outputDataFormat": "honeywell.quantum-results.v1", "outputDataUri":
         "https://mystorage.blob.core.windows.net/job-00000000-0000-0000-0000-000000000000/rawOutputData?sv=PLACEHOLDER&sr=b&sig=PLACEHOLDER&se=PLACEHOLDER&sp=r&rscd=attachment%3B%20filename%3Dquantinuum-job-00000000-0000-0000-0000-000000000000.output.json",
-        "creationTime": "2022-02-18T22:54:42.9672408+00:00", "beginExecutionTime":
-        "2022-02-18T22:54:44.575897+00:00", "endExecutionTime": "2022-02-18T22:54:44.576694+00:00",
+        "creationTime": "2022-02-19T01:14:56.3180959+00:00", "beginExecutionTime":
+        "2022-02-19T01:14:58.302875+00:00", "endExecutionTime": "2022-02-19T01:14:58.303435+00:00",
         "cancellationTime": null, "costEstimate": null, "errorData": null, "isCancelling":
         false, "tags": [], "access_token": "fake_token"}'
     headers:
       content-length:
-      - '1419'
+      - '1421'
       content-type:
       - application/json; charset=utf-8
       transfer-encoding:
@@ -293,7 +293,7 @@ interactions:
       User-Agent:
       - azsdk-python-storage-blob/12.9.0 Python/3.7.10 (Darwin-21.3.0-x86_64-i386-64bit)
       x-ms-date:
-      - Fri, 18 Feb 2022 22:54:48 GMT
+      - Sat, 19 Feb 2022 01:15:02 GMT
       x-ms-range:
       - bytes=0-33554431
       x-ms-version:
@@ -317,7 +317,7 @@ interactions:
       x-ms-blob-type:
       - BlockBlob
       x-ms-creation-time:
-      - Fri, 18 Feb 2022 22:54:43 GMT
+      - Sat, 19 Feb 2022 01:14:57 GMT
       x-ms-lease-state:
       - available
       x-ms-lease-status:
